+++ Magic/CHANGELOG.md
@@ -80,6 +80,7 @@
  - Magic attribute modifiers now support specific slots and operations like vanilla modifiers
  - Allow placeholders in Recall icon descriptions
  - Add `item_properties` wand config, for (possibly randomized) properties to be put on the wand item when it is created
+ - Add `hidden: true` option on attribute configs, will keep them out of wand lore
 
 ## Changes